+++ test/Internal/IRegistryHive.cs
@@ -1,0 +1,9 @@
+﻿using Microsoft.Win32;
+
+namespace ChromaWrapper.Tests.Internal
+{
+    internal interface IRegistryHive
+    {
+        RegistryKey Key { get; }
+    }
+}