+++ test/Internal/Xunit/XunitExtensions.cs
@@ -5,12 +5,19 @@ namespace ChromaWrapper.Tests.Internal.Xunit
 {
     internal static class XunitExtensions
     {
-        public static ITest? GetTest(this ITestOutputHelper output)
+        public static ITest GetTest(this ITestOutputHelper output)
         {
             // Reference: https://github.com/xunit/xunit/issues/416#issuecomment-378512739
-            return (ITest?)output.GetType()
-                .GetField("test", BindingFlags.Instance | BindingFlags.NonPublic)!
+            var res = (ITest?)output.GetType()
+                .GetField("test", BindingFlags.Instance | BindingFlags.NonPublic)?
                 .GetValue(output);
+
+            if (res == null)
+            {
+                throw new NotSupportedException();
+            }
+
+            return res;
         }
     }
 }